Automation: Added sql utils
Added execute and fetch sql util functions with gets a db connection
without considering if the connection is from postgresql or memsql
also added postgres and memsql wrapper utils
