+++ automation/devops_automation_infra/utils/sql.py
@@ -1,0 +1,14 @@
+
+
+
+def execute(connection, query):
+    with connection.cursor() as c:
+        res = c.execute(query)
+    connection.commit()
+    return res
+
+
+def fetchall(connection, query):
+    with connection.cursor() as c:
+        c.execute(query)
+        return c.fetchall()
